+++ raytracer/librt3/samplers/GridSampler.cs
@@ -71,7 +71,7 @@ namespace raytracer.samplers
             }
         }
 
-        public override float TotalSamples()
+        public override uint TotalSamples()
         {
             return Screen.Height*Screen.Width*NumberOfSamples;
         }
